Refactor(#188): 에러 컴포넌트 전역 적용
- 유저페이지의 에러 컴포넌트를 루트로 위치 이동
- 테스트 파일의 import 경로 수정
